+++ perceiver/model/audio/symbolic/backend.py
@@ -3,7 +3,12 @@ from dataclasses import dataclass, fields
 import torch
 import torch.nn as nn
 
-from perceiver.model.core import InputAdapter, OutputAdapter, PerceiverAR, PerceiverARConfig, positions, RotarySupport
+from perceiver.model.core import (
+    PerceiverAR,
+    PerceiverARConfig,
+    TiedTokenOutputAdapter,
+    TokenInputAdapterWithRotarySupport,
+)
 from perceiver.model.core.utils import init_parameters
 
 
@@ -23,72 +28,6 @@ class SymbolicAudioModelConfig(PerceiverARConfig):
         return cls(**{field.name: kwargs[field.name] for field in fields(cls) if field.name in kwargs})
 
 
-# TODO: move to common module (copied from TextInputAdapter)
-class SequenceInputAdapter(InputAdapter):
-    def __init__(self, vocab_size: int, max_seq_len: int, num_input_channels: int, abs_pos_emb: bool = True):
-        super().__init__(num_input_channels)
-        self._max_seq_len = max_seq_len
-        self._abs_pos_emb = abs_pos_emb
-
-        self.txt_embedding = nn.Embedding(vocab_size, num_input_channels)
-
-        if abs_pos_emb:
-            self.pos_embedding = nn.Embedding(max_seq_len, num_input_channels)
-
-    @property
-    def vocab_size(self):
-        return self.txt_embedding.num_embeddings
-
-    @property
-    def max_seq_len(self):
-        return self._max_seq_len
-
-    def forward(self, x, abs_pos=None):
-        if self._abs_pos_emb:
-            if abs_pos is None:
-                abs_pos = positions(*x.shape, device=x.device)
-            return self.txt_embedding(x) + self.pos_embedding(abs_pos)
-        else:
-            return self.txt_embedding(x)
-
-
-# TODO: move to common module (copied from TiedTextOutputAdapter)
-class TiedSequenceOutputAdapter(OutputAdapter):
-    def __init__(self, vocab_size: int, emb_bias: bool = True):
-        super().__init__()
-        self._emb_bias = emb_bias
-        if emb_bias:
-            self.bias = nn.Parameter(torch.zeros(vocab_size))
-
-    def forward(self, x, txt_embedding: nn.Embedding):
-        result = torch.matmul(x, txt_embedding.weight.T)
-        if self._emb_bias:
-            return result + self.bias
-        else:
-            return result
-
-
-class SymbolicAudioInputAdapter(RotarySupport, SequenceInputAdapter):
-    def __init__(
-        self,
-        rotated_channels_per_head: int,
-        vocab_size: int,
-        max_seq_len: int,
-        num_input_channels: int,
-        abs_pos_emb: bool,
-    ):
-        super().__init__(
-            rotated_channels_per_head=rotated_channels_per_head,
-            vocab_size=vocab_size,
-            max_seq_len=max_seq_len,
-            num_input_channels=num_input_channels,
-            abs_pos_emb=abs_pos_emb,
-        )
-
-    def forward(self, x, abs_pos=None):
-        return super().forward(x, abs_pos)
-
-
 # TODO: create common base class for SymbolicAudioModel and CausalLanguageModel
 class SymbolicAudioModel(PerceiverAR):
     def __init__(self, config: SymbolicAudioModelConfig):
@@ -98,7 +37,7 @@ class SymbolicAudioModel(PerceiverAR):
             # Rotary embedding only for first 50% of channels ...
             num_rotated_channels = num_rotated_channels // 2
 
-        input_adapter = SymbolicAudioInputAdapter(
+        input_adapter = TokenInputAdapterWithRotarySupport(
             rotated_channels_per_head=num_rotated_channels,
             vocab_size=config.vocab_size,
             max_seq_len=config.max_seq_len,
@@ -111,7 +50,7 @@ class SymbolicAudioModel(PerceiverAR):
         if config.output_norm:
             self.out_norm = nn.LayerNorm(config.num_channels)
 
-        self.output_adapter = TiedSequenceOutputAdapter(vocab_size=config.vocab_size, emb_bias=config.output_bias)
+        self.output_adapter = TiedTokenOutputAdapter(vocab_size=config.vocab_size, emb_bias=config.output_bias)
         self._init_parameters(config.init_scale)
 
     def _init_parameters(self, init_scale: float):
