+++ perceiver/model/core/adapter.py
@@ -81,3 +81,67 @@ class TrainableQueryProvider(nn.Module, QueryProvider):
 
     def forward(self, x=None):
         return rearrange(self._query, "... -> 1 ...")
+
+
+class TokenInputAdapter(InputAdapter):
+    def __init__(self, vocab_size: int, max_seq_len: int, num_input_channels: int, abs_pos_emb: bool = True):
+        super().__init__(num_input_channels)
+        self._max_seq_len = max_seq_len
+        self._abs_pos_emb = abs_pos_emb
+
+        self.txt_embedding = nn.Embedding(vocab_size, num_input_channels)
+
+        if abs_pos_emb:
+            self.pos_embedding = nn.Embedding(max_seq_len, num_input_channels)
+
+    @property
+    def vocab_size(self):
+        return self.txt_embedding.num_embeddings
+
+    @property
+    def max_seq_len(self):
+        return self._max_seq_len
+
+    def forward(self, x, abs_pos=None):
+        if self._abs_pos_emb:
+            if abs_pos is None:
+                abs_pos = positions(*x.shape, device=x.device)
+            return self.txt_embedding(x) + self.pos_embedding(abs_pos)
+        else:
+            return self.txt_embedding(x)
+
+
+class TokenInputAdapterWithRotarySupport(RotarySupport, TokenInputAdapter):
+    def __init__(
+        self,
+        rotated_channels_per_head: int,
+        vocab_size: int,
+        max_seq_len: int,
+        num_input_channels: int,
+        abs_pos_emb: bool,
+    ):
+        super().__init__(
+            rotated_channels_per_head=rotated_channels_per_head,
+            vocab_size=vocab_size,
+            max_seq_len=max_seq_len,
+            num_input_channels=num_input_channels,
+            abs_pos_emb=abs_pos_emb,
+        )
+
+    def forward(self, x, abs_pos=None):
+        return super().forward(x, abs_pos)
+
+
+class TiedTokenOutputAdapter(OutputAdapter):
+    def __init__(self, vocab_size: int, emb_bias: bool = True):
+        super().__init__()
+        self._emb_bias = emb_bias
+        if emb_bias:
+            self.bias = nn.Parameter(torch.zeros(vocab_size))
+
+    def forward(self, x, txt_embedding: nn.Embedding):
+        result = torch.matmul(x, txt_embedding.weight.T)
+        if self._emb_bias:
+            return result + self.bias
+        else:
+            return result
