+++ perceiver/model/text/clm/__init__.py
@@ -1,9 +1,4 @@
-from perceiver.model.text.clm.backend import (
-    CausalLanguageModel,
-    CausalLanguageModelConfig,
-    TextInputAdapter,
-    TextOutputAdapter,
-)
+from perceiver.model.text.clm.backend import CausalLanguageModel, CausalLanguageModelConfig
 from perceiver.model.text.clm.huggingface import (
     convert_checkpoint,
     PerceiverCausalLanguageModel,
