+++ perceiver/model/text/common/__init__.py
@@ -1,1 +1,1 @@
-from perceiver.model.text.common.backend import TextEncoder, TextEncoderConfig, TextInputAdapter, TiedTextOutputAdapter
+from perceiver.model.text.common.backend import TextEncoder, TextEncoderConfig
